Refactor space creation into a separate module
